+++ RELEASES.md
@@ -878,6 +878,7 @@
 -----------------
   - fix: SMC-3242 - Will correctly listRelevantPermissions for onBehalfOf sessions
 
+
 11.13.3 2021-09-08
 -----------------
   - fix: SMC-4044 - cookie events not dependent on client login
@@ -894,3 +895,12 @@
 11.13.6 2021-11-09
 -----------------
   - fix: SMC-4512 -  data provider get action causes fatal when allowNestedPermissions switched on
+
+
+11.14.0 2021-11-24
+-----------------
+  - feature: SMC-2954 - Allows for configuration and use of multiple authentication providers.
+
+11.14.1 2021-11-24
+-----------------
+  - fix: SMC-4386 -  upsertMultiplePermissions will now allow for removing permissions/prohibitions as well as upserting permissions or prohibitions